refactor: move tests to __tests__ to follow jest convention
