+++ jest.config.js
@@ -26,4 +26,9 @@ module.exports = {
   moduleNameMapper: {
     "src/(.*)": "<rootDir>/src/$1",
   },
+  modulePathIgnorePatterns: [
+    "__tests__/mocks",
+    "__tests__/fixtures",
+    "__tests__/utils",
+  ],
 };
