+++ src/__tests__/fixtures/flatfile/InitializeEmptyBatch.json
