+++ src/__tests__/fixtures/skylark/mutations/updateHomepageSetContent.json
