+++ src/__tests__/fixtures/skylark/queries/getObject/gotImage.json
