+++ src/__tests__/fixtures/skylark/queries/getObject/gots01e01.json
