+++ src/__tests__/fixtures/skylark/queries/getObject/homepage.json
