+++ src/__tests__/fixtures/skylark/queries/getObject/setWithContent.json
