+++ src/__tests__/fixtures/skylark/queries/introspection/objectTypes.json
