+++ src/__tests__/fixtures/skylark/queries/introspection/schema.json
