+++ src/__tests__/fixtures/skylark/queries/search/gotAssetsOnly.json
