+++ src/__tests__/mocks/handlers/updateObjectHandlers.ts
@@ -1,6 +1,6 @@
 import { graphql } from "msw";
 
-import GQLSkylarkGetSetWithContentQueryFixture from "src/tests/fixtures/skylark/queries/getObject/setWithContent.json";
+import GQLSkylarkGetSetWithContentQueryFixture from "src/__tests__/fixtures/skylark/queries/getObject/setWithContent.json";
 
 export const updateObjectHandlers = [
   graphql.mutation(`UPDATE_OBJECT_CONTENT_Set`, (req, res, ctx) => {
