+++ src/components/checkbox/checkbox.test.tsx
@@ -1,6 +1,6 @@
 import { fireEvent } from "@storybook/testing-library";
 
-import { render, screen } from "src/tests/utils/test-utils";
+import { render, screen } from "src/__tests__/utils/test-utils";
 
 import { Checkbox } from "./checkbox.component";
