+++ src/components/displayGraphQLQuery/graphQLQueryModal.test.tsx
@@ -1,5 +1,5 @@
+import { fireEvent, render, screen } from "src/__tests__/utils/test-utils";
 import { GET_SKYLARK_SCHEMA } from "src/lib/graphql/skylark/queries";
-import { fireEvent, render, screen } from "src/tests/utils/test-utils";
 
 import { DisplayGraphQLQuery } from "./graphQLQueryModal.component";
