+++ src/components/objectListing/objectListing.stories.tsx
@@ -2,6 +2,9 @@ import { ComponentStory } from "@storybook/react";
 import { userEvent, within } from "@storybook/testing-library";
 import { DocumentNode } from "graphql";
 
+import GQLSkylarkObjectTypesQueryFixture from "src/__tests__/fixtures/skylark/queries/introspection/objectTypes.json";
+import GQLSkylarkSchemaQueryFixture from "src/__tests__/fixtures/skylark/queries/introspection/schema.json";
+import GQLGameOfThronesSearchResults from "src/__tests__/fixtures/skylark/queries/search/got.json";
 import { SEARCH_PAGE_SIZE } from "src/hooks/useSearch";
 import { GQLSkylarkSchemaQueriesMutations } from "src/interfaces/graphql/introspection";
 import { createSearchObjectsQuery } from "src/lib/graphql/skylark/dynamicQueries";
@@ -10,9 +13,6 @@ import {
   GET_SKYLARK_SCHEMA,
 } from "src/lib/graphql/skylark/queries";
 import { getAllObjectsMeta } from "src/lib/skylark/objects";
-import GQLSkylarkObjectTypesQueryFixture from "src/tests/fixtures/skylark/queries/introspection/objectTypes.json";
-import GQLSkylarkSchemaQueryFixture from "src/tests/fixtures/skylark/queries/introspection/schema.json";
-import GQLGameOfThronesSearchResults from "src/tests/fixtures/skylark/queries/search/got.json";
 
 import { ObjectList } from "./objectListing.component";
