+++ src/components/objectListing/search/searchInput/searchInput.test.tsx
@@ -1,7 +1,7 @@
 import { fireEvent } from "@storybook/testing-library";
 import { act } from "react-dom/test-utils";
 
-import { render, screen } from "src/tests/utils/test-utils";
+import { render, screen } from "src/__tests__/utils/test-utils";
 
 import { SearchInput } from "./searchInput.component";
