+++ src/components/user/avatar/avatar.test.tsx
@@ -1,7 +1,7 @@
 import "@testing-library/jest-dom";
 import { act } from "react-dom/test-utils";
 
-import { render, screen } from "src/tests/utils/test-utils";
+import { render, screen } from "src/__tests__/utils/test-utils";
 
 import { UserAvatar } from "./avatar.component";
