+++ src/lib/graphql/skylark/dynamicMutations.test.ts
@@ -1,12 +1,12 @@
 import {
-  ParsedSkylarkObjectConfig,
-  ParsedSkylarkObjectMetadata,
-} from "src/interfaces/skylark";
-import {
   episodeObjectOperations,
   setObjectOperations,
   movieObjectOperations,
-} from "src/tests/utils/objectOperations";
+} from "src/__tests__/utils/objectOperations";
+import {
+  ParsedSkylarkObjectConfig,
+  ParsedSkylarkObjectMetadata,
+} from "src/interfaces/skylark";
 
 import {
   createDeleteObjectMutation,
