+++ src/lib/skylark/introspection/introspection.test.ts
@@ -1,12 +1,12 @@
 import { GraphQLClient } from "graphql-request";
 import { MockApolloClient, createMockClient } from "mock-apollo-client";
 
+import GQLSkylarkObjectTypesQueryFixture from "src/__tests__/fixtures/skylark/queries/introspection/objectTypes.json";
 import { createSkylarkClient } from "src/lib/graphql/skylark/client";
 import {
   GET_SKYLARK_OBJECT_TYPES,
   GET_SKYLARK_SCHEMA,
 } from "src/lib/graphql/skylark/queries";
-import GQLSkylarkObjectTypesQueryFixture from "src/tests/fixtures/skylark/queries/introspection/objectTypes.json";
 
 import {
   getSkylarkObjectOperations,
